+++ Applications/Custom/LayerPlugin/layer_plugin_test.cpp
@@ -19,7 +19,6 @@
 
 #include <app_context.h>
 #include <layer.h>
-#include <layer_internal.h>
 
 const char *NNTRAINER_PATH = std::getenv("NNTRAINER_PATH");
