+++ debian/nntrainer-dev.install
@@ -13,7 +13,7 @@
 /usr/include/nntrainer/graph_node.h
 /usr/include/nntrainer/databuffer.h
 /usr/include/nntrainer/databuffer_factory.h
-/usr/include/nntrainer/layer_internal.h
+/usr/include/nntrainer/layer_devel.h
 /usr/include/nntrainer/neuralnet.h
 /usr/include/nntrainer/tensor.h
 /usr/include/nntrainer/tensor_dim.h
