+++ nntrainer/layers/activation_layer.cpp
@@ -20,7 +20,6 @@
 
 #include <activation_layer.h>
 #include <blas_interface.h>
-#include <layer_internal.h>
 #include <lazy_tensor.h>
 #include <nntrainer_error.h>
 #include <nntrainer_log.h>
@@ -73,9 +72,9 @@ void ActivationLayer::setProperty(const std::vector<std::string> &values) {
 
 void ActivationLayer::setProperty(const std::string &type_str,
                                   const std::string &value) {
-  using PropertyType = LayerV1::PropertyType;
-  LayerV1::PropertyType type =
-    static_cast<LayerV1::PropertyType>(parseLayerProperty(type_str));
+  using PropertyType = nntrainer::Layer::PropertyType;
+  nntrainer::Layer::PropertyType type =
+    static_cast<nntrainer::Layer::PropertyType>(parseLayerProperty(type_str));
 
   switch (type) {
   case PropertyType::activation: {
