+++ nntrainer/layers/bn_layer.cpp
@@ -22,7 +22,6 @@
  */
 
 #include <bn_layer.h>
-#include <layer_internal.h>
 #include <lazy_tensor.h>
 #include <nntrainer_error.h>
 #include <nntrainer_log.h>
@@ -102,10 +101,10 @@ void BatchNormalizationLayer::setProperty(
 
 void BatchNormalizationLayer::setProperty(const std::string &type_str,
                                           const std::string &value) {
-  using PropertyType = LayerV1::PropertyType;
+  using PropertyType = nntrainer::Layer::PropertyType;
   int status = ML_ERROR_NONE;
-  LayerV1::PropertyType type =
-    static_cast<LayerV1::PropertyType>(parseLayerProperty(type_str));
+  nntrainer::Layer::PropertyType type =
+    static_cast<nntrainer::Layer::PropertyType>(parseLayerProperty(type_str));
 
   switch (type) {
   case PropertyType::epsilon:
