+++ nntrainer/layers/concat_layer.cpp
@@ -13,7 +13,6 @@
 
 #include <concat_layer.h>
 #include <cstring>
-#include <layer_internal.h>
 #include <nntrainer_error.h>
 #include <nntrainer_log.h>
 #include <parse_util.h>
