+++ nntrainer/layers/conv2d_layer.cpp
@@ -18,7 +18,6 @@
 
 #include <blas_interface.h>
 #include <conv2d_layer.h>
-#include <layer_internal.h>
 #include <lazy_tensor.h>
 #include <nntrainer_error.h>
 #include <nntrainer_log.h>
@@ -502,10 +501,10 @@ void Conv2DLayer::setProperty(const std::vector<std::string> &values) {
 
 void Conv2DLayer::setProperty(const std::string &type_str,
                               const std::string &value) {
-  using PropertyType = LayerV1::PropertyType;
+  using PropertyType = nntrainer::Layer::PropertyType;
   int status = ML_ERROR_NONE;
-  LayerV1::PropertyType type =
-    static_cast<LayerV1::PropertyType>(parseLayerProperty(type_str));
+  nntrainer::Layer::PropertyType type =
+    static_cast<nntrainer::Layer::PropertyType>(parseLayerProperty(type_str));
 
   switch (type) {
   case PropertyType::filters: {
