+++ nntrainer/layers/input_layer.cpp
@@ -22,7 +22,6 @@
  */
 
 #include <input_layer.h>
-#include <layer_internal.h>
 #include <nntrainer_error.h>
 #include <nntrainer_log.h>
 #include <parse_util.h>
@@ -54,10 +53,10 @@ void InputLayer::setProperty(const std::vector<std::string> &values) {
 
 void InputLayer::setProperty(const std::string &type_str,
                              const std::string &value) {
-  using PropertyType = LayerV1::PropertyType;
+  using PropertyType = nntrainer::Layer::PropertyType;
   int status = ML_ERROR_NONE;
-  LayerV1::PropertyType type =
-    static_cast<LayerV1::PropertyType>(parseLayerProperty(type_str));
+  nntrainer::Layer::PropertyType type =
+    static_cast<nntrainer::Layer::PropertyType>(parseLayerProperty(type_str));
 
   switch (type) {
   case PropertyType::normalization: {
