+++ nntrainer/layers/lstm.cpp
@@ -12,11 +12,11 @@
  */
 
 #include <cmath>
-#include <layer_internal.h>
 #include <lazy_tensor.h>
 #include <lstm.h>
 #include <nntrainer_error.h>
 #include <nntrainer_log.h>
+#include <node_exporter.h>
 #include <parse_util.h>
 #include <util_func.h>
 
@@ -132,10 +132,10 @@ void LSTMLayer::setProperty(const std::vector<std::string> &values) {
 
 void LSTMLayer::setProperty(const std::string &type_str,
                             const std::string &value) {
-  using PropertyType = LayerV1::PropertyType;
+  using PropertyType = nntrainer::Layer::PropertyType;
   int status = ML_ERROR_NONE;
-  LayerV1::PropertyType type =
-    static_cast<LayerV1::PropertyType>(parseLayerProperty(type_str));
+  nntrainer::Layer::PropertyType type =
+    static_cast<nntrainer::Layer::PropertyType>(parseLayerProperty(type_str));
 
   // TODO : Add return_state property & api to get the hidden input
   switch (type) {
