+++ nntrainer/layers/output_layer.cpp
@@ -12,7 +12,6 @@
  */
 
 #include <cstring>
-#include <layer_internal.h>
 #include <nntrainer_error.h>
 #include <nntrainer_log.h>
 #include <output_layer.h>
