+++ nntrainer/layers/plugged_layer.h
@@ -36,7 +36,7 @@ public:
     layerImpl(pluggable->createfunc()),
     destroy_func(pluggable->destroyfunc) {
     NNTR_THROW_IF(layerImpl == nullptr, std::invalid_argument)
-      << "either create_func_ failed or cannot dynamic cast to layer_internal";
+      << "either create_func_ failed or cannot dynamic cast to layer";
   }
 
   /**
