+++ nntrainer/layers/preprocess_translate_layer.cpp
@@ -14,7 +14,6 @@
 
 #include <random>
 
-#include <layer_internal.h>
 #include <nntrainer_error.h>
 #include <nntrainer_log.h>
 #include <parse_util.h>
@@ -82,10 +81,10 @@ void PreprocessTranslateLayer::setProperty(
 
 void PreprocessTranslateLayer::setProperty(const std::string &type_str,
                                            const std::string &value) {
-  using PropertyType = LayerV1::PropertyType;
+  using PropertyType = nntrainer::Layer::PropertyType;
   int status = ML_ERROR_NONE;
-  LayerV1::PropertyType type =
-    static_cast<LayerV1::PropertyType>(parseLayerProperty(type_str));
+  nntrainer::Layer::PropertyType type =
+    static_cast<nntrainer::Layer::PropertyType>(parseLayerProperty(type_str));
 
   switch (type) {
   case PropertyType::random_translate: {
