+++ nntrainer/layers/split_layer.cpp
@@ -12,7 +12,6 @@
  */
 
 #include <cstring>
-#include <layer_internal.h>
 #include <nntrainer_error.h>
 #include <nntrainer_log.h>
 #include <parse_util.h>
@@ -171,10 +170,10 @@ void SplitLayer::setProperty(const std::vector<std::string> &values) {
 
 void SplitLayer::setProperty(const std::string &type_str,
                              const std::string &value) {
-  using PropertyType = LayerV1::PropertyType;
+  using PropertyType = nntrainer::Layer::PropertyType;
   int status = ML_ERROR_NONE;
-  LayerV1::PropertyType type =
-    static_cast<LayerV1::PropertyType>(parseLayerProperty(type_str));
+  nntrainer::Layer::PropertyType type =
+    static_cast<nntrainer::Layer::PropertyType>(parseLayerProperty(type_str));
 
   switch (type) {
   case PropertyType::split_dimension: {
