+++ nntrainer/layers/tflite_layer.cpp
@@ -10,7 +10,6 @@
  * @bug    No known bugs except for NYI items
  */
 
-#include <layer_internal.h>
 #include <nntrainer_error.h>
 #include <nntrainer_log.h>
 #include <tflite_layer.h>
@@ -94,9 +93,9 @@ void TfLiteLayer::setProperty(const std::vector<std::string> &values) {
 
 void TfLiteLayer::setProperty(const std::string &type_str,
                               const std::string &value) {
-  using PropertyType = LayerV1::PropertyType;
-  LayerV1::PropertyType type =
-    static_cast<LayerV1::PropertyType>(parseLayerProperty(type_str));
+  using PropertyType = nntrainer::Layer::PropertyType;
+  nntrainer::Layer::PropertyType type =
+    static_cast<nntrainer::Layer::PropertyType>(parseLayerProperty(type_str));
 
   switch (type) {
   case PropertyType::modelfile: {
