+++ nntrainer/layers/time_dist.cpp
@@ -11,7 +11,6 @@
  *
  */
 
-#include <layer_internal.h>
 #include <nntrainer_error.h>
 #include <nntrainer_log.h>
 #include <parse_util.h>
